# docs/openapi-spec.yml
openapi: 3.0.3
info:
  title: 찰떡 서비스 API
  description: 찰떡 서비스의 전체 REST API 명세 (자동 생성, 주요 기능별 한국어 설명 포함)
  version: 1.0.0
servers:
  - url: /api/v1
    description: 기본 API 서버
paths:
...
  /shops/{id}:
    get:
      summary: 매장 상세 조회
      description: 매장 상세 정보 조회
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
          description: 매장 ID
      responses:
        '200':
          description: 매장 정보 반환
        '404':
          description: 매장 없음
  # 리뷰
  /reviews/shops/{shopId}:
    get:
      summary: 매장 리뷰 목록 조회
      description: 특정 매장의 리뷰 목록 조회
      parameters:
        - in: path
          name: shopId
          required: true
          schema:
            type: integer
          description: 매장 ID
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: 리뷰 목록 반환
    post:
      summary: 리뷰 작성
      description: JWT 인증 필요, 리뷰 작성
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: shopId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: 리뷰 작성 성공
  /reviews/{reviewId}:
    get:
      summary: 리뷰 상세 조회
      description: 리뷰 상세 정보 조회
      parameters:
        - in: path
          name: reviewId
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 리뷰 정보 반환
        '404':
          description: 리뷰 없음
  /shops/{shopId}/reviews/{reviewId}:
    put:
      summary: 리뷰 수정
      description: JWT 인증 필요, 리뷰 수정
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: shopId
          required: true
          schema:
            type: integer
        - in: path
          name: reviewId
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdate'
      responses:
        '200':
          description: 리뷰 수정 성공
  # 신고
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ReviewCreate:
      type: object
      properties:
        description:
          type: string
        shopPrdId:
          type: integer
    ReviewUpdate:
      type: object
      properties:
        description:
          type: string
...